Simplify the example. now, main branch contains examples, other than real system. branch "SMARTLINK" is continous be ok for smartlink project

+++ ZoomPipeline_FuncSvr/smartlink/st_clientnode_applayer.cpp
@@ -3,16 +3,13 @@
 #include <QSqlDatabase>
 #include <QSqlQuery>
 #include <QSqlError>
-namespace SmartLink{
+namespace ExampleServer{
 	st_clientNodeAppLayer::st_clientNodeAppLayer(st_client_table * pClientTable, QObject * pClientSock ,QObject *parent) :
 		st_clientNode_baseTrans(pClientTable,pClientSock,parent)
 	{
 
 		m_bLoggedIn= false;
-		memset(&m_current_app_header,0,sizeof(SMARTLINK_MSG_APP));
-		memset(m_serialNum,0,sizeof(m_serialNum));
-		memset(m_username,0,sizeof(m_username));
-
+		memset(&m_current_app_header,0,sizeof(EXAMPLE_MSG_APP));
 	}
 
 	bool st_clientNodeAppLayer::loadRelations()
@@ -20,16 +17,8 @@ namespace SmartLink{
 		QSqlDatabase db = m_pClientTable->dbRes()->databse(m_pClientTable->Database_UserAcct());
 		if (db.isValid()==true && db.isOpen()==true )
 		{
-			QString sql = "select ";
-			if (m_serialNum[0]!=0)
-				sql += "user_id from relations where equip_id = ?;";
-			else if (m_username[0]!=0)
-				sql += "equip_id from relations where user_id = ?;";
-			else
-			{
-				emit evt_Message(this,tr("try to save relations before login!"));
-				return false;
-			}
+			QString sql = "select friend_id from relations where user_id = ?;";
+
 			QSqlQuery query(db);
 			query.prepare(sql);
 			query.addBindValue((quint32)m_uuid);
@@ -60,60 +49,31 @@ namespace SmartLink{
 		QSqlDatabase db = m_pClientTable->dbRes()->databse(m_pClientTable->Database_UserAcct());
 		if (db.isValid()==true && db.isOpen()==true )
 		{
-			QString sql = "delete from relations where ";
-			if (m_serialNum[0]!=0)
+			QSqlQuery query(db);
+			QString sql = "delete from relations where user_id = ?;";
+
+			query.prepare(sql);
+			query.addBindValue((quint32)m_uuid);
+			if (false== query.exec())
 			{
-				QSqlQuery query(db);
-				sql += "equip_id = ?;";
-				query.prepare(sql);
-				query.addBindValue((quint32)m_uuid);
-				if (false== query.exec())
-				{
-					emit evt_Message(this,tr("try to del old relations Failed! ")+ query.lastError().text());
-					return false;
-				}
-			}
-			else if (m_username[0]!=0)
-			{
-				QSqlQuery query(db);
-				sql += "user_id = ?;";
-				query.prepare(sql);
-				query.addBindValue((quint32)m_uuid);
-				if (false== query.exec())
-				{
-					emit evt_Message(this,tr("try to del old relations Failed! ")+ query.lastError().text());
-					return false;
-				}
-			}
-			else
-			{
-				emit evt_Message(this,tr("try to save relations before login!"));
+				emit evt_Message(this,tr("try to del old relations Failed! ")+ query.lastError().text());
 				return false;
 			}
 
-			sql = "insert into relations (equip_id,user_id) values (?,?);";
+
+
+			sql = "insert into relations (user_id,friend_id) values (?,?);";
 			//the forigen key can automatic avoid non-existing values.
 			foreach(quint32 nodeid,m_matched_nodes)
 			{
 
 				QSqlQuery query(db);
-				if (m_serialNum[0]!=0)
-				{
-					if (bIsValidUserId(nodeid)==false)
-						continue;
-					query.prepare(sql);
-					query.addBindValue((quint32)m_uuid);
-					query.addBindValue(nodeid);
+				if (bIsValidUserId(nodeid)==false)
+					continue;
+				query.prepare(sql);
+				query.addBindValue((quint32)m_uuid);
+				query.addBindValue(nodeid);
 
-				}
-				else if (m_username[0]!=0)
-				{
-					if (bIsValidEquipId(nodeid)==false)
-						continue;
-					query.prepare(sql);
-					query.addBindValue(nodeid);
-					query.addBindValue((quint32)m_uuid);
-				}
 				if (false== query.exec())
 				{
 					emit evt_Message(this,tr("try to insert new relations Failed! ")+ query.lastError().text());
@@ -153,17 +113,7 @@ namespace SmartLink{
 			}
 			else
 			{
-				if (bIsValidEquipId(m_currentHeader.source_id))
-				{
-					//Deal Box->Svr Msgs
-					if (false==Deal_Node2Svr_Msgs())
-					{
-						m_currentBlock = QByteArray();
-						emit evt_Message(this,tr("Box To Server Message Failed."));
-						emit evt_close_client(this->sock());
-					}
-				}
-				else if (bIsValidUserId(m_currentHeader.source_id) )
+				if (bIsValidUserId(m_currentHeader.source_id) )
 				{
 					//Deal Client->Svr Msgs
 					if (false==Deal_Node2Svr_Msgs())
@@ -187,21 +137,6 @@ namespace SmartLink{
 				}
 			}
 
-		}
-		//deal Broadcast messages
-		else if (m_currentHeader.destin_id==0xFFFFFFFC)
-		{
-			//need furture works.
-			//Do Nothing
-			emit evt_Message(this,tr("Broadcast Message is not currently supported."));
-			m_currentBlock = QByteArray();
-		}
-		else if (m_currentHeader.destin_id==0xFFFFFFFD)
-		{
-			//need furture works.
-			//Do Nothing
-			emit evt_Message(this,tr("Broadcast Message is not currently supported."));
-			m_currentBlock = QByteArray();
 		}
 		else
 		{
@@ -237,9 +172,9 @@ namespace SmartLink{
 		bool res = true;
 		//qDebug()<<m_currentHeader.data_length<<"\n";
 		//qDebug()<<this->m_currentBlock.toHex()<<"\n";
-		if (m_currentHeader.data_length < sizeof (SMARTLINK_MSG_APP::tag_app_layer_header))
+		if (m_currentHeader.data_length < sizeof (EXAMPLE_MSG_APP::tag_app_layer_header))
 			return false;
-		if (m_currentMessageSize < sizeof(SMARTLINK_MSG) - 1 + sizeof (SMARTLINK_MSG_APP::tag_app_layer_header))
+		if (m_currentMessageSize < sizeof(EXAMPLE_TRANS_MSG) - 1 + sizeof (EXAMPLE_MSG_APP::tag_app_layer_header))
 		{
 			// header is not complete, return
 			return true;
@@ -247,49 +182,19 @@ namespace SmartLink{
 		//Catch the header
 		if (m_current_app_header.header.MsgType==0x00)
 			memcpy((void *)&this->m_current_app_header,
-				   ((unsigned char *)this->m_currentBlock.constData()) + sizeof(SMARTLINK_MSG) - 1,
-				   sizeof (SMARTLINK_MSG_APP::tag_app_layer_header)
+				   ((unsigned char *)this->m_currentBlock.constData()) + sizeof(EXAMPLE_TRANS_MSG) - 1,
+				   sizeof (EXAMPLE_MSG_APP::tag_app_layer_header)
 				   );
 		//qDebug()<<m_current_app_header.header.MsgType<<"\n";
 		switch (m_current_app_header.header.MsgType)
 		{
-		case 0x1000:
-			if (bytesLeft()>0)
-				// message is not complete, return
-				return true;
-			if (m_currentMessageSize>
-					sizeof(SMARTLINK_MSG) - 1
-					+ sizeof (SMARTLINK_MSG_APP::tag_app_layer_header)
-					+ sizeof (stMsg_HostRegistReq)+64)
-			{
-				emit evt_Message(this,tr("Broken Message stMsg_HostRegistReq, size not correct."));
-				res = false;
-			}
-			else
-				res = this->RegisitNewBoxNode();
-			break;
-		case 0x1001:
-			if (bytesLeft()>0)
-				// message is not complete, return
-				return true;
-			if (m_currentMessageSize>
-					sizeof(SMARTLINK_MSG) - 1
-					+ sizeof (SMARTLINK_MSG_APP::tag_app_layer_header)
-					+ sizeof (stMsg_HostLogonReq)+64)
-			{
-				emit evt_Message(this,tr("Broken Message stMsg_HostLogonReq, size not correct."));
-				res = false;
-			}
-			else
-				res = this->LoginBox();
-			break;
 		case 0x3000:
 			if (bytesLeft()>0)
 				// message is not complete, return
 				return true;
 			if (m_currentMessageSize>
-					sizeof(SMARTLINK_MSG) - 1
-					+ sizeof (SMARTLINK_MSG_APP::tag_app_layer_header)
+					sizeof(EXAMPLE_TRANS_MSG) - 1
+					+ sizeof (EXAMPLE_MSG_APP::tag_app_layer_header)
 					+ sizeof (stMsg_ClientLoginReq)+66)
 			{
 				emit evt_Message(this,tr("Broken Message stMsg_ClientLoginReq, size not correct."));
@@ -315,9 +220,9 @@ namespace SmartLink{
 	{
 		bool res = true;
 
-		if (m_currentHeader.data_length < sizeof (SMARTLINK_MSG_APP::tag_app_layer_header))
+		if (m_currentHeader.data_length < sizeof (EXAMPLE_MSG_APP::tag_app_layer_header))
 			return false;
-		if (m_currentMessageSize < sizeof(SMARTLINK_MSG) - 1 + sizeof (SMARTLINK_MSG_APP::tag_app_layer_header))
+		if (m_currentMessageSize < sizeof(EXAMPLE_TRANS_MSG) - 1 + sizeof (EXAMPLE_MSG_APP::tag_app_layer_header))
 		{
 			// header is not complete, return
 			return true;
@@ -325,34 +230,20 @@ namespace SmartLink{
 		//Catch the header
 		if (m_current_app_header.header.MsgType==0x00)
 			memcpy((void *)&this->m_current_app_header,
-				   ((unsigned char *)this->m_currentBlock.constData()) + sizeof(SMARTLINK_MSG) - 1,
-				   sizeof (SMARTLINK_MSG_APP::tag_app_layer_header)
+				   ((unsigned char *)this->m_currentBlock.constData()) + sizeof(EXAMPLE_TRANS_MSG) - 1,
+				   sizeof (EXAMPLE_MSG_APP::tag_app_layer_header)
 				   );
 		//do
 		switch (m_current_app_header.header.MsgType)
 		{
-		case 0x1002:
-			if (bytesLeft()>0)
-				// message is not complete, return
-				return true;
-			if (m_currentMessageSize!=
-					sizeof(SMARTLINK_MSG) - 1
-					+ sizeof (SMARTLINK_MSG_APP::tag_app_layer_header)
-					/*+ sizeof (stMsg_HostTimeCorrectReq)*/)
-			{
-				emit evt_Message(this,tr("Broken Message stMsg_HostRegistReq, size not correct."));
-				res = false;
-			}
-			else
-				res = this->Box2Svr_CorrectTime();
-			break;
+
 		case 0x1003:
 			if (bytesLeft()>0)
 				return true;
 
 			if (m_currentMessageSize<
-					sizeof(SMARTLINK_MSG) - 1
-					+ sizeof (SMARTLINK_MSG_APP::tag_app_layer_header)
+					sizeof(EXAMPLE_TRANS_MSG) - 1
+					+ sizeof (EXAMPLE_MSG_APP::tag_app_layer_header)
 					+ sizeof (stMsg_UploadUserListReq) - sizeof(quint32))
 			{
 				emit evt_Message(this,tr("Broken Message stMsg_UploadUserListReq, size not correct."));
@@ -366,8 +257,8 @@ namespace SmartLink{
 				// message is not complete, return
 				return true;
 			if (m_currentMessageSize!=
-					sizeof(SMARTLINK_MSG) - 1
-					+ sizeof (SMARTLINK_MSG_APP::tag_app_layer_header)
+					sizeof(EXAMPLE_TRANS_MSG) - 1
+					+ sizeof (EXAMPLE_MSG_APP::tag_app_layer_header)
 					/*+ sizeof (stMsg_DownloadUserListReq)*/)
 			{
 				emit evt_Message(this,tr("Broken Message stMsg_DownloadUserListReq, size not correct."));
@@ -381,8 +272,8 @@ namespace SmartLink{
 				// message is not complete, return
 				return true;
 			if (m_currentMessageSize!=
-					sizeof(SMARTLINK_MSG) - 1
-					+ sizeof (SMARTLINK_MSG_APP::tag_app_layer_header)
+					sizeof(EXAMPLE_TRANS_MSG) - 1
+					+ sizeof (EXAMPLE_MSG_APP::tag_app_layer_header)
 					+ sizeof (stMsg_ClientLogoutReq))
 			{
 				emit evt_Message(this,tr("Broken Message stMsg_ClientLogoutReq, size not correct."));
@@ -391,21 +282,7 @@ namespace SmartLink{
 			else
 				res = this->ClientLogout();
 			break;
-		case 0x3002:
-			if (bytesLeft()>0)
-				// message is not complete, return
-				return true;
-			if (m_currentMessageSize!=
-					sizeof(SMARTLINK_MSG) - 1
-					+ sizeof (SMARTLINK_MSG_APP::tag_app_layer_header)
-					/*+ sizeof (stMsg_GetHostListReq)*/)
-			{
-				emit evt_Message(this,tr("Broken Message stMsg_GetHostListReq, size not correct."));
-				res = false;
-			}
-			else
-				res = this->GetHostList();
-			break;
+
 		default:
 			emit evt_Message(this,tr("Message type not supported."));
 			res = false;
